Move tests to safety_tests app
