+++ migrate.py
@@ -3,4 +3,4 @@ from django.core.management import call_command
 from boot_django import boot_django
 
 boot_django()
-call_command("migrate", "safety")
+call_command("migrate")
