+++ pyproject.toml
@@ -12,8 +12,9 @@ packages = [{ include = "safety" }]
 include = ["LICENSE"]
 
 [tool.poetry.dependencies]
-python = ">3.5"
+python = ">=3.7"
 django = ">3.2"
+django-fake-model = "^0.1.4"
 
 [tool.poetry.group.dev.dependencies]
 build = "^0.10.0"
