+++ safety_tests/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class SafetyTestsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'safety_tests'
